+++ radiantcore/rendersystem/backend/OpenGLShader.h
@@ -56,7 +56,7 @@ private:
     void constructLightingPassesFromMaterial();
     void determineBlendModeForEditorPass(OpenGLState& pass);
     void constructEditorPreviewPassFromMaterial();
-    void applyAlphaTestToPass(OpenGLState& pass, float alphaTest);
+    void applyAlphaTestToPass(OpenGLState& pass, double alphaTest);
     void setGLTexturesFromTriplet(OpenGLState&, const DBSTriplet&);
 
     // Destroy internal data
